+++ mycodo/outputs/pcf8574.py
@@ -170,7 +170,12 @@ class OutputModule(AbstractOutput):
 
         for channel in channels_dict:
             if self.options_channels['trigger_functions_startup'][channel]:
-                self.check_triggers(self.unique_id, output_channel=channel)
+                try:
+                    self.check_triggers(self.unique_id, output_channel=channel)
+                except Exception as err:
+                    self.logger.error(
+                        "Could not check Trigger for channel {} of output {}: {}".format(
+                            channel, self.unique_id, err))
 
     def output_switch(self,
                       state,